+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Compatibility with TYPO3 v11 LTS
+- Variable resolver for page properties
 
 ### Removed
 - Compatibility with PHP 7.2